+++ src/vramsteg.cpp
@@ -109,15 +109,15 @@ int main (int argc, char** argv)
     signal (SIGUSR1,   SIG_IGN);
     signal (SIGUSR2,   SIG_IGN);
 
-    int         arg_current    = 0;
+    long        arg_current    = 0;
 #ifdef WAITING_FOR_VITAPI
     std::string arg_done       = "";
 #endif
     bool        arg_elapsed    = false;
     bool        arg_estimate   = false;
     std::string arg_label;
-    int         arg_max        = 0;
-    int         arg_min        = 0;
+    long        arg_max        = 0;
+    long        arg_min        = 0;
     bool        arg_percentage = false;
 #ifdef WAITING_FOR_VITAPI
     std::string arg_remaining  = "";
@@ -165,15 +165,15 @@ int main (int argc, char** argv)
     {
       switch (ch)
       {
-      case 'c': arg_current    = atoi (optarg);        break;
+      case 'c': arg_current    = atol (optarg);        break;
 #ifdef WAITING_FOR_VITAPI
       case 'd': arg_done       = optarg;               break;
 #endif
       case 'e': arg_elapsed    = true;                 break;
       case 't': arg_estimate   = true;                 break;
       case 'l': arg_label      = optarg;               break;
-      case 'x': arg_max        = atoi (optarg);        break;
-      case 'm': arg_min        = atoi (optarg);        break;
+      case 'x': arg_max        = atol (optarg);        break;
+      case 'm': arg_min        = atol (optarg);        break;
       case 'n': std::cout << time (NULL) << std::endl; exit (0);
       case 'p': arg_percentage = true;                 break;
 #ifdef WAITING_FOR_VITAPI
